cms/requirements.txt
asgiref==3.8.1
chardet==5.2.0
cssselect2==0.7.0
Django==4.2
django-admin-rangefilter==0.13.2
django-filer==3.2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-polymorphic==3.1.0
easy-thumbnails==2.10
lxml==5.3.0
pillow==10.4.0
psycopg2-binary==2.9.9
reportlab==4.2.4
setuptools==75.1.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
tzdata==2024.2
webencodings==0.5.1
